usuarios completado

+++ ProjectBackEnd/src/main/java/com/crud/h2/dao/IUsuarioDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crud.h2.dto.Usuario;
 
-public interface IUsuarioDAO extends JpaRepository<Usuario, Long>{
+public interface IUsuarioDAO extends JpaRepository<Usuario, Integer>{
 	
 		public List<Usuario> findByNombre(String nombre);
